Add README

+++ df-init.sh
@@ -1,18 +1,20 @@
-# Instructions:
-#
-#     rm .bash_aliases .vimrc
-#     curl --remote-name https://raw.githubusercontent.com/NedJWestern/dotfiles/master/dfinit.txt
-#     mv dfinit.txt dfinit.sh
-#     chmod +x dfinit.sh
-#     ./dfinit.sh
-#     rm dfinit.sh
+# exit on error
+set -e
 
+if [[ -f ~/.bash_aliases ]] || [[ -f ~/.vimrc ]]; then
+    echo ERROR: Please first remove or backup these files: .bash_aliases, .vimrc
+    exit 1
+fi
+
+# prevent git recursion issues
 echo ".dotfiles" > $HOME/.gitignore
+
+# clone and setup git repo
 git clone --bare https://github.com/NedJWestern/dotfiles.git $HOME/.dotfiles
 git --git-dir=$HOME/.dotfiles/ --work-tree=$HOME checkout
 git --git-dir=$HOME/.dotfiles/ --work-tree=$HOME config --local status.showUntrackedFiles no
 
-# Add source bash aliases
+# Add sourcing of bash aliases
 if ! grep -q ".bash_aliases" ~/.bashrc; then
 
     cat << 'EOT' >> ~/.bashrc
@@ -24,5 +26,3 @@ fi
 EOT
 fi
 
-
-source .bashrc
